backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosqlite==0.19.0
websockets>=12.0
bcrypt>=4.0.1
psutil>=5.9.0
